RavenDB-5244 V4.0 client long running task - Split AsyncDocumentSession to partial classes
